+++ Core/Util/DisArm64.cpp
@@ -120,14 +120,14 @@ void DecodeBitMasks(int immN, int imms, int immr, uint64_t *tmask, uint64_t *wma
 	// if len < 1 then ReservedValue();
 	// assert M >= (1 << len);
 	// Determine S, R and S - R parameters
-	int levels = Ones(len);
+	int levels = (int)Ones(len);
 	uint32_t S = imms & levels;
 	uint32_t R = immr & levels;
 	int diff = S - R; // 6-bit subtract with borrow
 	int esize = 1 << len;
 	int d = diff & Ones(len - 1);
-	uint32_t welem = Ones(S + 1);
-	uint32_t telem = Ones(d + 1);
+	uint32_t welem = (uint32_t)Ones(S + 1);
+	uint32_t telem = (uint32_t)Ones(d + 1);
 	if (wmask) {
 		uint64_t rotated = ROR(welem, R, esize);
 		*wmask = Replicate(rotated, esize);
